switch type to size_t as there is a signed vs unsigned compare, and calloc takes a size_t type

+++ src/workorder.h
@@ -153,7 +153,7 @@ int workorder_num_stats(workorder_t *);
 int workorder_num_strands_bytype(workorder_t *, uint32_t);
 int group_max_open_connections(group_t *);
 int workorder_num_connections(workorder_t *);
-int group_max_dto_size(group_t *);
+size_t group_max_dto_size(group_t *);
 void group_free(group_t *);
 group_t* group_clone(group_t *);
 int group_opposite(group_t *);
